Refactor into smaller files for easier readability
