+++ src/controller/index.js
@@ -1,11 +1,11 @@
 import { Botkit } from "botkit";
-import { receiveMiddleware, sendMiddleware } from "./middleware";
 import { adapter } from "./slack-adapter";
+import * as middleware from "./middleware";
 
 export const controller = new Botkit({
   webhook_uri: "/api/messages",
   adapter: adapter,
 });
 
-controller.middleware.receive.use(receiveMiddleware);
-controller.middleware.send.use(sendMiddleware);
+controller.middleware.receive.use(middleware.receive);
+controller.middleware.send.use(middleware.send);
