+++ src/controller/slack-adapter.js
